+++ bindings/react/src/components/ListItem.jsx
@@ -27,6 +27,10 @@ class ListItem extends SimpleWrapper {
   componentDidMount() {
     super.componentDidMount();
     this.node = ReactDOM.findDOMNode(this);
+    this.node.expanded = this.props.expanded === true;
+    if (this.node.expanded) {
+      this.node.classList.add('expanded');
+    }
   }
 
   componentDidUpdate() {